enhancement(editor): extracting simple functions to helpers module
